+++ Cargo.toml
@@ -20,6 +20,10 @@ fnv = "1.0"
 
 [features]
 render_wiring_steps = []
+pcb-naive = []
+pcb-hashmap = []
+pcb-grid = []
+default = ["pcb-grid"]
 
 [profile.release]
 debug = true
